# Change stack variable from vector to hashmap