--- a/crates/nu-protocol/src/engine/capture_block.rs
+++ b/crates/nu-protocol/src/engine/capture_block.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use crate::{BlockId, Value, VarId};
 
 use serde::{Deserialize, Serialize};
@@ -5,5 +7,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Closure {
     pub block_id: BlockId,
-    pub captures: Vec<(VarId, Value)>,
+    pub captures: Box<HashMap<VarId, Value>>,
 }