--- a/crates/nu-protocol/src/engine/stack.rs
+++ b/crates/nu-protocol/src/engine/stack.rs
@@ -34,7 +34,7 @@ pub type EnvVars = HashMap<String, HashMap<String, Value>>;
 #[derive(Debug, Clone)]
 pub struct Stack {
     /// Variables
-    pub vars: Vec<(VarId, Value)>,
+    pub vars: Box<HashMap<VarId, Value>>,
     /// Environment variables arranged as a stack to be able to recover values from parent scopes
     pub env_vars: Vec<Arc<EnvVars>>,
     /// Tells which environment variables from engine state are hidden, per overlay.
@@ -72,7 +72,7 @@ impl Stack {
     /// (as opposed to a [`PipelineData`](crate::PipelineData)).
     pub fn new() -> Self {
         Self {
-            vars: Vec::new(),
+            vars: Box::new(HashMap::new()),
             env_vars: Vec::new(),
             env_hidden: Arc::new(HashMap::new()),
             active_overlays: vec![DEFAULT_OVERLAY_NAME.to_string()],
@@ -101,7 +101,7 @@ impl Stack {
             error_handlers: ErrorHandlerStack::new(),
             use_ir: parent.use_ir,
             recursion_count: parent.recursion_count,
-            vars: vec![],
+            vars: Box::new(HashMap::default()),
             parent_deletions: vec![],
             config: parent.config.clone(),
             out_dest: parent.out_dest.clone(),
@@ -123,8 +123,9 @@ impl Stack {
 
         unique_stack
             .vars
-            .retain(|(var, _)| !child.parent_deletions.contains(var));
-        for (var, value) in child.vars {
+            .retain(|var, _| !child.parent_deletions.contains(var));
+
+        for (var, value) in child.vars.into_iter() {
             unique_stack.add_var(var, value);
         }
         unique_stack.env_vars = child.env_vars;
@@ -151,7 +152,7 @@ impl Stack {
 
     /// Lookup a variable, returning None if it is not present
     fn lookup_var(&self, var_id: VarId) -> Option<Value> {
-        for (id, val) in &self.vars {
+        for (id, val) in self.vars.iter() {
             if var_id == *id {
                 return Some(val.clone());
             }
@@ -229,22 +230,16 @@ impl Stack {
 
     pub fn add_var(&mut self, var_id: VarId, value: Value) {
         //self.vars.insert(var_id, value);
-        for (id, val) in &mut self.vars {
-            if *id == var_id {
-                *val = value;
-                return;
+        match self.vars.get_mut(&var_id) {
+            Some(val) => *val = value,
+            None => {
+                self.vars.insert(var_id, value);
             }
         }
-        self.vars.push((var_id, value));
     }
 
     pub fn remove_var(&mut self, var_id: VarId) {
-        for (idx, (id, _)) in self.vars.iter().enumerate() {
-            if *id == var_id {
-                self.vars.remove(idx);
-                break;
-            }
-        }
+        self.vars.remove(&var_id);
         // even if we did have it in the original layer, we need to make sure to remove it here
         // as well (since the previous update might have simply hid the parent value)
         if self.parent_stack.is_some() {
@@ -300,12 +295,15 @@ impl Stack {
             })
     }
 
-    pub fn captures_to_stack(&self, captures: Vec<(VarId, Value)>) -> Stack {
+    pub fn captures_to_stack(&self, captures: Box<HashMap<VarId, Value>>) -> Stack {
         self.captures_to_stack_preserve_out_dest(captures)
             .collect_value()
     }
 
-    pub fn captures_to_stack_preserve_out_dest(&self, captures: Vec<(VarId, Value)>) -> Stack {
+    pub fn captures_to_stack_preserve_out_dest(
+        &self,
+        captures: Box<HashMap<VarId, Value>>,
+    ) -> Stack {
         let mut env_vars = self.env_vars.clone();
         env_vars.push(Arc::new(HashMap::new()));
 
@@ -326,7 +324,7 @@ impl Stack {
     }
 
     pub fn gather_captures(&self, engine_state: &EngineState, captures: &[VarId]) -> Stack {
-        let mut vars = Vec::with_capacity(captures.len());
+        let mut vars = Box::new(HashMap::with_capacity(captures.len()));
 
         let fake_span = Span::new(0, 0);
 
@@ -334,9 +332,9 @@ impl Stack {
             // Note: this assumes we have calculated captures correctly and that commands
             // that take in a var decl will manually set this into scope when running the blocks
             if let Ok(value) = self.get_var(*capture, fake_span) {
-                vars.push((*capture, value));
+                vars.insert(*capture, value);
             } else if let Some(const_val) = &engine_state.get_var(*capture).const_val {
-                vars.push((*capture, const_val.clone()));
+                vars.insert(*capture, const_val.clone());
             }
         }