# Description
This PR updates the variable stack in the Nu engine to use a HashMap. Previously, searching, adding, or removing a variable required looping through the entire vector, which was inefficient. By switching to a HashMap, we improve time complexity. The following tables show the benchmark between the old and new versions using [Hyperfine](https://github.com/sharkdp/hyperfine) on the release mode.
| Script |  Old | New |
| -------- | -------- | -------- |
| [fibonaci-recursive](https://github.com/nushell/nu_scripts/blob/main/benchmarks/fibonacci-recursive.nu) | 3.431 s ±  0.150 s     | 3.313 s ±  0.129 s |
| [fibonaci](https://github.com/nushell/nu_scripts/blob/main/benchmarks/fibonacci.nu)               |854.7 ms ±  37.7 ms     |  818.5 ms ±   7.6 ms |
| [gradient_benchmark_no_check](https://github.com/nushell/nu_scripts/blob/main/benchmarks/gradient_benchmark_no_check.nu) |251.6 ms ±   3.2 ms s | 237.8 ms ±   3.1 ms |
| [turtle](https://github.com/nushell/nu_scripts/blob/main/benchmarks/turtle.nu)                  |25.0 ms ±   0.8 ms     | 24.7 ms ±   0.8 ms |

I believe we can further optimize by switching from the standard HashMap to [FxHashMap](https://github.com/cbreeden/fxhash). If this PR looks good, I can create a follow-up PR.