@sholderbach 2024-10-07
Tried a few variations in just the `Stack` and can't quite reproduce your numbers:

![grafik](https://github.com/user-attachments/assets/f6f8e545-702b-40b8-a566-b559d0ad9db0)

https://github.com/sholderbach/nushell/tree/hashnu