@sholderbach 2024-10-08
Thanks for trying and attempting to benchmark it! This takes patience with our build times alone.

Our codebase is by no means optimized yet and it is always valuable to have a fresh set of eyes on the key components.